requirements.txt
streamlit==1.28.0
torch==2.0.1
torchvision==0.15.2
transformers==4.33.0
pillow==10.0.0
python-dotenv==1.0.0
requests==2.31.0
opencv-python-headless==4.8.0.76
ultralytics==8.0.124
colorthief==0.2.1
numpy==1.24.3
huggingface_hub==0.16.4
plotly==5.15.0
scikit-learn==1.3.0

